third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-object-element/object-attributes-expected.txt

This is a testharness.js-based test.
[FAIL] object.contentWindow
  assert_equals: The contentWindow of the object element should be null as it is showing fallback content. expected null but got object "[object Window]"
Harness: the test ran to completion.
